chore: ::load() path now saved for snapshots
