+++ src/snapshot/mod.rs
@@ -1,4 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod state;
-pub use state::{Snapshot, get_snapshot_path};
+pub use state::Snapshot;
+pub mod path;
+pub use path::get_snapshot_path;
